function that sets the value of a bit to 1 at a given index.

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -8,7 +8,7 @@
  */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned long int count;
+	unsigned int count;
 
 	if (n == NULL)
 		return (-1);
